requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   openai
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
distro==1.9.0
    # via openai
et-xmlfile==2.0.0
    # via openpyxl
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
jiter==0.6.1
    # via openai
numpy==2.1.2
    # via
    #   -r requirements.in
    #   pandas
openai==1.52.2
    # via -r requirements.in
openpyxl==3.1.5
    # via -r requirements.in
pandas==2.2.3
    # via -r requirements.in
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   openai
pydantic-core==2.23.4
    # via pydantic
pymupdf==1.24.13
    # via -r requirements.in
pypdf2==3.0.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tenacity==9.0.0
    # via -r requirements.in
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   openai
    #   pydantic
    #   pydantic-core
tzdata==2024.2
    # via pandas
